Fixed FacebookUserData.timezone type from int to double. Due ot Mumbai timezone 5.5

+++ Oauth2Login/Service/FacebookService.cs
@@ -76,7 +76,7 @@ namespace Oauth2Login.Service
         public string gender { get; set; }
         public string picture { get; set; }
         public string locale { get; set; }
-        public int timezone { get; set; }
+        public double timezone { get; set; }
         public bool verified { get; set; }
 
         // override
